Factor out SortTask to fix shadows
