+++ libraries/render-utils/src/RenderShadowTask.cpp
@@ -9,15 +9,19 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
+#include "RenderShadowTask.h"
+
 #include <gpu/Context.h>
 
 #include <ViewFrustum.h>
 
-#include "render/Context.h"
+#include <render/Context.h>
+#include <render/CullTask.h>
+#include <render/SortTask.h>
+#include <render/DrawTask.h>
+
 #include "DeferredLightingEffect.h"
 #include "FramebufferCache.h"
-
-#include "RenderShadowTask.h"
 
 #include "model_shadow_vert.h"
 #include "skin_model_shadow_vert.h"
@@ -28,7 +32,7 @@
 using namespace render;
 
 void RenderShadowMap::run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext,
-                          const render::ShapesIDsBounds& inShapes) {
+                          const render::ShapeBounds& inShapes) {
     assert(renderContext->args);
     assert(renderContext->args->_viewFrustum);
 
@@ -111,11 +115,11 @@ RenderShadowTask::RenderShadowTask(CullFunctor cullFunctor) : Task(std::make_sha
     const auto culledShadowSelection = addJob<CullSpatialSelection>("CullShadowSelection", shadowSelection, cullFunctor, RenderDetails::SHADOW, shadowFilter);
 
     // Sort
-    const auto sortedShapes = addJob<PipelineSortShapes>("PipelineSortShadowSort", culledShadowSelection);
-    const auto shadowShapes = addJob<DepthSortItems>("DepthSortShadowMap", sortedShapes);
+    const auto sortedPipelines = addJob<PipelineSortShapes>("PipelineSortShadowSort", culledShadowSelection);
+    const auto sortedShapes = addJob<DepthSortShapes>("DepthSortShadowMap", sortedPipelines);
 
     // GPU jobs: Render to shadow map
-    addJob<RenderShadowMap>("RenderShadowMap", shadowShapes, shapePlumber);
+    addJob<RenderShadowMap>("RenderShadowMap", sortedShapes, shapePlumber);
 }
 
 void RenderShadowTask::configure(const Config& configuration) {
